requirements.txt
google-generativeai>=0.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pillow>=10.0.0
mcp>=0.1.0
proto>=1.0.0 